Refactor and fix DB access
