+++ libs/block-mesh-manager/src/ws/mod.rs
@@ -1,4 +1,6 @@
+pub mod broadcaster;
 pub mod connection_manager;
+pub mod cron_reports_controller;
 pub mod handle_socket;
 pub mod process_message;
 pub mod task_scheduler;
