+++ libs/extension/extension_js/manifests/manifest_cr.json
@@ -4,7 +4,7 @@
   "name": "BlockMesh Network",
   "short_name": "BlockMesh",
   "description": "BlockMesh Network allows you to generate passive income from your excessive bandwidth",
-  "version": "0.0.250",
+  "version": "0.0.252",
   "author": "ohaddahan@blockmesh.xyz",
   "minimum_chrome_version": "110",
   "externally_connectable": {
